backend/requirements.txt
# Backend dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
httpx==0.27.0
sqlalchemy==2.0.27
alembic==1.13.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0
authlib==1.2.1
psycopg2-binary==2.9.9  # For PostgreSQL